Fix bug with job 10

+++ utils/steps-v3/preprocessing/initialize_cancel_script.sh
@@ -19,12 +19,16 @@ cat utils/kill_session.sh >> ${kill_ssh}
 
 
 # KILL_SH: File that runs on the user space
-clean_jobs="/pw/jobs/${workflow_name}/*$((job_number_int-10))"
+job_number_to_clean=$((job_number_int-10))
+clean_jobs="/pw/jobs/${workflow_name}/${job_number_to_clean}"
 
 echo "#!/bin/bash" > ${kill_sh}
 echo "mv ${kill_sh} ${kill_sh}.completed" >> ${kill_sh}
 cat resources/host/inputs.sh >> ${kill_sh}
-echo "trap \"rm -rf ${clean_jobs}\" EXIT" >> ${kill_sh}
+if [ "${job_number_to_clean}" -gt 0 ]; then
+    echo "trap \"rm -rf ${clean_jobs}\" EXIT" >> ${kill_sh}
+fi
+
 echo "echo Running ${kill_sh}" >> ${kill_sh}
 # Add kill_ssh
 cat >> ${kill_sh} <<HERE
